requirements.txt
# Code Analysis
ast-comments>=1.0.0
radon>=5.1.0
pylint>=2.15.0

# AI/ML
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Graph & Visualization
networkx>=3.0
matplotlib>=3.7.0
graphviz>=0.20.1
plotly>=5.14.0

# File Handling
GitPython>=3.1.31
pathspec>=0.11.1

# Web Interface (Optional)
streamlit>=1.22.0
flask>=2.3.0

# Reporting
markdown>=3.4.3
jinja2>=3.1.2
python-docx>=0.8.11

# Utilities
pyyaml>=6.0
colorlog>=6.7.0
tqdm>=4.65.0
requests>=2.31.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0